REF: pandas/core/window.py into multiple files (#27736)
