+++ pandas/core/window/__init__.py
@@ -1,0 +1,3 @@
+from pandas.core.window.ewm import EWM  # noqa:F401
+from pandas.core.window.expanding import Expanding, ExpandingGroupby  # noqa:F401
+from pandas.core.window.rolling import Rolling, RollingGroupby, Window  # noqa:F401
