+++ pandas/tests/window/test_window.py
@@ -6,7 +6,7 @@ import pandas.util._test_decorators as td
 
 import pandas as pd
 from pandas import Series
-import pandas.core.window as rwindow
+from pandas.core.window import Window
 from pandas.tests.window.common import Base
 
 
@@ -50,7 +50,7 @@ class TestWindow(Base):
     @pytest.mark.parametrize("method", ["sum", "mean"])
     def test_numpy_compat(self, method):
         # see gh-12811
-        w = rwindow.Window(Series([2, 4, 6]), window=[0, 2])
+        w = Window(Series([2, 4, 6]), window=[0, 2])
 
         msg = "numpy operations are not valid with window objects"
